requirements.txt
flask==3.0.3
requests==2.32.3
numpy==1.26.4
tensorflow==2.17.0
werkzeug==3.0.4
python-dotenv==1.0.1
flask-limiter==3.8.0
gunicorn==21.2.0
gevent==23.9.1
